Fix invitation URL handling by removing initialLocation
The issue was that GoRouter's initialLocation parameter was set to '/', which was overriding the actual URL that users were trying to visit. When users visited /invite/SDIMUW7G, the app would start at '/' instead of the invitation URL.

By removing the initialLocation parameter, GoRouter will now use the actual URL that the user is visiting, allowing invitation links to work correctly.

This should fix the invitation flow where users were being redirected to the home screen instead of seeing the invitation preview.